#97 fix sonar

+++ saturn-console-api/src/main/java/com/vip/saturn/job/console/domain/JobStatistics.java
@@ -127,7 +127,7 @@ public class JobStatistics implements Serializable {
 		if (processCountOfAllTime == 0) {
 			return 0;
 		}
-		float rate = (float) errorCountOfAllTime / processCountOfAllTime;
+		double rate = (double) errorCountOfAllTime / processCountOfAllTime;
 		return (float) (Math.floor(rate * 10000) / 10000.0);
 	}
